# StringOperator: Improve performance of Equals()