… src/Quartz.Tests.Unit/Serialized/StringOperator_Anything.ser
… src/Quartz.Tests.Unit/Serialized/StringOperator_Contains.ser
… src/Quartz.Tests.Unit/Serialized/StringOperator_EndsWith.ser
… src/Quartz.Tests.Unit/Serialized/StringOperator_Equality.ser
… src/Quartz.Tests.Unit/Serialized/StringOperator_StartsWith.ser