Speed up `Equals(...)` by first performing a reference comparison.
Update docs.
Add (de)serialization tests.

This is a clear improvement for .NET Framework, and a minor regression (within noise range) for .NET 5.0.
As such, I ifdef'ed out this optimalisation for .NET 5.0 and higher.

<details>
<summary>Benchmark results</summary>

|                              Method | CLR                | Branch    |      Mean |     Error |    StdDev |
|------------------------------------ | ------------------:|----------:|----------:|----------:|----------:|
|         Equals_Object_SameReference | .NET Framework 4.8 | main      |  7.617 ns | 0.0701 ns | 0.0656 ns |
|         Equals_Object_SameReference | .NET Framework 4.8 | PR        | 2.2835 ns | 0.0719 ns | 0.0706 ns |
|         Equals_Object_SameReference |           .NET 5.0 | main      | 1.9304 ns | 0.0099 ns | 0.0093 ns |
|         Equals_Object_SameReference |           .NET 5.0 | PR        | 1.9173 ns | 0.0046 ns | 0.0041 ns |
|              Equals_Object_NotEqual | .NET Framework 4.8 | main      |  7.590 ns | 0.0291 ns | 0.0243 ns |
|              Equals_Object_NotEqual | .NET Framework 4.8 | PR        | 7.5199 ns | 0.0093 ns | 0.0083 ns |
|              Equals_Object_NotEqual |           .NET 5.0 | main      | 0.7403 ns | 0.0047 ns | 0.0044 ns |
|              Equals_Object_NotEqual |           .NET 5.0 | PR        | 0.8407 ns | 0.0028 ns | 0.0025 ns |
| Equals_StringOperator_SameReference | .NET Framework 4.8 | main      |  5.544 ns | 0.0186 ns | 0.0164 ns |
| Equals_StringOperator_SameReference | .NET Framework 4.8 | PR        | 0.5449 ns | 0.0091 ns | 0.0085 ns |
| Equals_StringOperator_SameReference |           .NET 5.0 | main      | 0.2283 ns | 0.0037 ns | 0.0031 ns |
| Equals_StringOperator_SameReference |           .NET 5.0 | PR        | 0.4729 ns | 0.0008 ns | 0.0007 ns |
|      Equals_StringOperator_NotEqual | .NET Framework 4.8 | main      |  5.579 ns | 0.0823 ns | 0.0770 ns |
|      Equals_StringOperator_NotEqual | .NET Framework 4.8 | PR        | 5.5100 ns | 0.0055 ns | 0.0051 ns |
|      Equals_StringOperator_NotEqual |           .NET 5.0 | main      | 0.6217 ns | 0.0059 ns | 0.0055 ns |
|      Equals_StringOperator_NotEqual |           .NET 5.0 | PR        | 0.8275 ns | 0.0022 ns | 0.0020 ns |

</details>